fix: separate components
